Tidy up material module
Move material.py into __init__.py and frontload the actual material
creation. Some docs were added to the MaterialCreator.
